python/requirements.txt
freetype-py==2.5.1
pillow==12.1.1
